@@ -10,9 +10,9 @@ version '1.5.0'
 repositories {
     jcenter()
     mavenCentral()
-    maven { url = 'https://papermc.io/repo/repository/maven-public/' }
-    maven { url = 'https://rayzr.dev/repo' }
-    maven { url = 'https://libraries.minecraft.net' }
+    maven { url = 'https://repo.papermc.io/repository/maven-public/' }
+    maven { url = 'https://rayzr.dev/repo/' }
+    maven { url = 'https://libraries.minecraft.net/' }
 }
 
 // PlotSquared API
@@ -26,7 +26,7 @@ dependencies {
 
 java {
     toolchain {
-        languageVersion = JavaLanguageVersion.of(16)
+        languageVersion = JavaLanguageVersion.of(17)
     }
 }
 
